Добавлены взрывы комнат

+++ Armlet/Armlet3/BastiliaGate_sw/Server/HonorLogic/ShipStatus/LakShip.cs
@@ -32,6 +32,6 @@ namespace HonorLogic.ShipStatus
 
         public string Name { get; set; }
         public IGlobalModel Model { get; set; }
-        public List<int> ShipRoomsIDs { get; set; }
+        public List<byte> ShipRoomsIDs { get; set; }
     }
 }
