+++ Armlet/Armlet3/BastiliaGate_sw/Server/PillInterfaces/IShip.cs
@@ -9,7 +9,7 @@ namespace HonorInterfaces
         bool SetSubsystemStatus(ShipSubsystemStatus subsystemStatus);
         List<ShipSubsystemStatus> GetAllSubsystemsStatus();
 
-        List<int> ShipRoomsIDs { get; set; }
+        List<byte> ShipRoomsIDs { get; set; }
 
         Guid ShipGuid { get; set; }
         int[] PhysicalGateID { get; set; }
